+++ WebUI/PreferencesUI/SkyDisplayPreferences.m
@@ -510,8 +510,8 @@ static NSNumber *NoNumber  = nil;
 }
 
 - (id)addPageAtTop {
-  unsigned int idx;
-  id           page;
+  NSUInteger idx;
+  id         page;
 
   page = [[[self item] retain] autorelease];
   idx  = [self->undockedPages indexOfObjectIdenticalTo:page];
